Renamed Facade to Decorator for Path, Directory and merge File facade with decorator
It was overkill having "Facade" and "Decorator" classes
Decorator alone is enough for now
