+++ CsharpExtras/Api/CsharpExtrasApi.cs
@@ -73,11 +73,9 @@ namespace CsharpExtras.Api
             return NewOneBasedArray(zeroBased);
         }
 
-        public IFileDecorator NewFileDecorator() => new FileDecoratorImpl(NewFileFacade());
-
-        public IFileFacade NewFileFacade() => new FileFacadeImpl();
-        public IDirectoryFacade NewDirectoryFacade() => new DirectoryFacadeImpl();
-        public IPathFacade NewPathFacade() => new PathFacadeImpl();
+        public IFileDecorator NewFileDecorator() => new FileDecoratorImpl();
+        public IDirectoryDecorator NewDirectoryDecorator() => new DirectoryDecoratorImpl();
+        public IPathDecorator NewPathDecorator() => new PathDecoratorImpl();
         public ILazyFunctionMap<TKey, TVal> NewLazyFunctionMap<TKey, TVal>(Func<TKey, TVal> backingFunction)
             => new LazyFunctionMapImpl<TKey, TVal>(backingFunction);
         public IRandomStringGenerator NewRandomStringGenerator() => new RandomStringGeneratorImpl();
