+++ CsharpExtras/IO/FileDecoratorImpl.cs
@@ -1,5 +1,6 @@
 ﻿using CsharpExtras.IO.FileNameCheck;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
@@ -7,13 +8,35 @@ namespace CsharpExtras.IO
 {
     class FileDecoratorImpl : IFileDecorator
     {
-        public IFileFacade FileFacade { get; }
-
         public IFileNameChecker FileNameChecker { get; set; } = new SystemFileNameChecker();
-
-        public FileDecoratorImpl(IFileFacade fileFacade)
+        public void Create(string path)
         {
-            FileFacade = fileFacade;
+            File.Create(path);
+        }
+
+        public string[] ReadAllLines(string path)
+        {
+            return File.ReadAllLines(path);
+        }
+
+        public string ReadAllText(string path)
+        {
+            return File.ReadAllText(path);
+        }
+
+        public void WriteAllLines(string path, IEnumerable<string> contents)
+        {
+            File.WriteAllLines(path, contents);
+        }
+
+        public void WriteAllText(string path, string contents)
+        {
+            File.WriteAllText(path, contents);
+        }
+
+        public bool Exists(string path)
+        {
+            return File.Exists(path);
         }
 
         public void TrimEmptyLinesFromEndOfFile(string filePath)
@@ -23,7 +46,7 @@ namespace CsharpExtras.IO
 
         public void TrimEmptyLinesFromEndOfFile(string filePath, Func<string, bool> isLineEmpty)
         {
-            string[] lines = FileFacade.ReadAllLines(filePath);
+            string[] lines = ReadAllLines(filePath);
             int maxIndex = lines.Count();
 
             for (int i = lines.Count() - 1; i > 0; i--)
@@ -35,7 +58,7 @@ namespace CsharpExtras.IO
                 maxIndex--;
             }
 
-            FileFacade.WriteAllLines(filePath, lines.Take(maxIndex));
+            WriteAllLines(filePath, lines.Take(maxIndex));
         }
 
         public bool IsValidFileName(string fileName)
