--- CsharpExtras/IO/FileFacadeImpl.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.IO;
-
-namespace CsharpExtras.IO
-{
-    class FileFacadeImpl : IFileFacade
-    {
-        public void Create(string path)
-        {
-            File.Create(path);
-        }
-
-        public string[] ReadAllLines(string path)
-        {
-            return File.ReadAllLines(path);
-        }
-
-        public string ReadAllText(string path)
-        {
-            return File.ReadAllText(path);
-        }
-
-        public void WriteAllLines(string path, IEnumerable<string> contents)
-        {
-            File.WriteAllLines(path, contents);
-        }
-
-        public void WriteAllText(string path, string contents)
-        {
-            File.WriteAllText(path, contents);
-        }
-
-        public bool Exists(string path)
-        {
-            return File.Exists(path);
-        }
-    }
-}
