--- CsharpExtras/IO/PathFacadeImpl.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CsharpExtras.IO
-{
-    class PathFacadeImpl : IPathFacade
-    {
-        public string GetExtension(string path)
-        {
-            return Path.GetExtension(path);
-        }
-
-        public string GetFileName(string path)
-        {
-            return Path.GetFileName(path);
-        }
-
-        public string GetDirectoryName(string path)
-        {
-            return Path.GetDirectoryName(path);
-        }
-
-
-        public string GetFileNameWithoutExtension(string path)
-        {
-            return Path.GetFileNameWithoutExtension(path);
-        }
-    }
-}
